feat(tx report): default endDate to end of current day
